fix1

+++ app/src/main/java/com/lh/zksocketc/utils/HttpUtil.java
@@ -1,14 +1,15 @@
 package com.lh.zksocketc.utils;
 
+import com.google.gson.Gson;
+import com.google.gson.reflect.TypeToken;
 import com.lh.zksocketc.MyApplication;
 import com.lh.zksocketc.data.DbDao.IcCardNumerDao;
+import com.lh.zksocketc.data.model.HttpData;
+import com.lh.zksocketc.data.model.HttpRow;
 import com.lh.zksocketc.data.model.IcCardNumer;
 
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-
 import java.io.IOException;
+import java.util.List;
 
 import okhttp3.Call;
 import okhttp3.Callback;
@@ -23,8 +24,7 @@ public class HttpUtil {
         OkHttpClient okHttpClient = new OkHttpClient();
         //2.创建Request对象，设置一个url地址（百度地址）,设置请求方式。
         Request request = new Request.Builder()
-                .url("https://smart.rovemaker.com/api/school/menu")
-//                .url(url)
+                .url(url + "/api/get_ic_card_list")
                 .build();
         //3.创建一个call对象,参数就是Request请求对象
         Call call = okHttpClient.newCall(request);
@@ -41,20 +41,22 @@ public class HttpUtil {
             public void onResponse(Call call, Response response) throws IOException {
                 String responseText = response.body().string();
                 ELog.e("==========数据=======" + responseText);
-                try {
-                    JSONObject jsonObject = new JSONObject(responseText);
-                    JSONArray jsonArray = jsonObject.getJSONArray("data");
+                Gson gson = new Gson();
+                HttpData httpData = gson.fromJson(responseText, HttpData.class);
+                ELog.e("==========IcCard==response=====" + httpData.toString());
+                if (httpData.flag == 1) {
+                    HttpData<HttpRow<List<IcCardNumer>>> httpRowHttpData = gson.fromJson(responseText, new TypeToken<HttpData<HttpRow<List<IcCardNumer>>>>() {
+                    }.getType());
+                    ELog.e("========IcCard==get=====" + httpRowHttpData);
                     IcCardNumerDao icCardNumerDao = MyApplication.getDaoSession().getIcCardNumerDao();
                     icCardNumerDao.deleteAll();
-                    for (int i = 0; i < jsonArray.length(); i++) {
-                        String IcCardNumer = jsonArray.getJSONObject(i).get("value").toString();
-                        icCardNumerDao.insert(new IcCardNumer(IcCardNumer));
+                    for (int i = 0; i < httpRowHttpData.getData().getRows().size(); i++) {
+                        icCardNumerDao.update(httpRowHttpData.getData().getRows().get(i));
                     }
-                } catch (JSONException e) {
-                    e.printStackTrace();
+
+                } else {
+
                 }
-
-
             }
         });
